finally

+++ src/main/java/org/beauty/tablebook/controllers/bookings/BookingRequest.java
@@ -2,12 +2,13 @@ package org.beauty.tablebook.controllers.bookings;
 
 import lombok.Getter;
 
+import java.time.LocalDateTime;
 import java.util.Date;
 @Getter
 public class BookingRequest {
 
     private Long restaurantId;
-    private Long tableId;
+    private Integer tableId;
     private Long userId;
     private Date time;
 }
